requirements.txt
dbt-postgres==0.20
dbt-snowflake==0.20
apache-airflow==1.10.14
wtforms==2.3.3
apache-airflow[postgres]==1.10.14
dbt-core==0.20
markupsafe==2.0.1
SQLAlchemy==1.3.23
flask